Add files via upload
This is the latest version of the affirmations program that centers the text on the screen followed by a drawing of a tree. It currently works on the BBC BASIC for Z80 version 1.06 included with the Agon Light 2 only. It does not work very well with BBC BASIC V yet. It appears it has to do with the system variables I am using to calculate the width of the screen and such.